testing/web-platform/tests/resource-timing/buffer-full-decrease-buffer-during-callback.html

<!DOCTYPE HTML>
<html>
<head onload>
<meta charset="utf-8" />
<title>This test validates that decreasing the buffer size in onresourcetimingbufferfull callback does not result in extra entries being dropped.</title>
<link rel="author" title="Intel" href="http://www.intel.com/" />
<link rel="help" href="http://www.w3.org/TR/resource-timing/#performanceresourcetiming"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/append.js"></script>
</head>
<body>
<script>
    const resource_timing_buffer_size = 2;
    const t = async_test("Verify that reducing the size of the resource timing buffer during resourcetimingbufferfull call works");
    performance.clearResourceTimings();
    let resize = function() {
        performance.setResourceTimingBufferSize(resource_timing_buffer_size - 1);
    }
    performance.addEventListener('resourcetimingbufferfull', resize);
    performance.setResourceTimingBufferSize(resource_timing_buffer_size);
    // Scripts appended in JS to ensure setResourceTimingBufferSize is called before.
    // This resource gets buffered in the resource timing entry buffer.
    xhrScript('resources/empty.js');
    xhrScript('resources/empty.js?second');
    // This resource overflows the entry buffer, and goes into the secondary buffer.
    xhrScript('resources/empty_script.js');
    window.onload = t.step_func_done(()=>{
        let entries = performance.getEntriesByType('resource');
        assert_equals(entries.length, 2,
                      'Both entries should be stored in resource timing buffer since it decreased its limit only after it overflowed.');
        assert_true(entries[0].name.includes('empty.js'), "empty.js is in the entries buffer");
        assert_true(entries[1].name.includes('empty.js?second'), "empty.js?second is in the entries buffer");
    });
</script>
</body>
</html>
